// app/Interfaces.ts
export interface LinkedTower{
  ad_tower: string
  pulse_building:string
}

export interface PropertyStatistics{
  interval: string; 
  sale_transaction: number; 
  max_price:number; 
  avg_price:number; 
  min_price: number ; 
  max_price_sqft:number; 
  avg_price_sqft: number; 
  min_price_sqft: number; 
  current_vs_avg_perc: number
}